// app/common/interfaces.ts
export interface IRendererLog {
  /**
   * Name of the window that sent the message.
   */
  wname: string;
  /**
   * Level of the message sent by the renderer.
   */
  type: "info" | "warn" | "error";
  /**
   * Message sent by the renderer process.
   */
  message: string;
}